+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Conventional Commits](https://www.conventionalcommi
 
 - Dependencies.
 - Examples - `bevy_ecs_tilemap` now supports `bevy_remote`.
+- Use `bevy::platform::collections::HashMap` instead of `std::collections::HashMap`.
 
 ## [0.8.0] - 2025-04-xx: Update to Bevy 0.16 and Edition 2024